refactored to remove storyboard, now need to fix sizing
