+++ business.xcodeproj/project.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:/Users/gustavomunhoz/Documents/Apps/Business/business/business.xcodeproj">
+      location = "self:">
    </FileRef>
 </Workspace>
