+++ business/Common/AppDelegate.swift
@@ -14,7 +14,16 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        // Override point for customization after application launch.
+        window = UIWindow(frame: UIScreen.main.bounds)
+        
+        let startViewController = MenuViewController()
+        
+        let navigationController = UINavigationController(rootViewController: startViewController)
+        
+        window?.rootViewController = navigationController
+        
+        window?.makeKeyAndVisible()
+        
         return true
     }
