+++ business/Info.plist
@@ -3,8 +3,25 @@
 <plist version="1.0">
 <dict>
 	<key>UIAppFonts</key>
-	<array>
-		<string></string>
-	</array>
+    <array>
+        <string>Urbanist-Black.ttf</string>
+        <string>Urbanist-BlackItalic.ttf</string>
+        <string>Urbanist-Bold.ttf</string>
+        <string>Urbanist-BoldItalic.ttf</string>
+        <string>Urbanist-ExtraBold.ttf</string>
+        <string>Urbanist-ExtraBoldItalic.ttf</string>
+        <string>Urbanist-ExtraLight.ttf</string>
+        <string>Urbanist-ExtraLightItalic.ttf</string>
+        <string>Urbanist-Italic.ttf</string>
+        <string>Urbanist-Light.ttf</string>
+        <string>Urbanist-LightItalic.ttf</string>
+        <string>Urbanist-Medium.ttf</string>
+        <string>Urbanist-MediumItalic.ttf</string>
+        <string>Urbanist-Regular.ttf</string>
+        <string>Urbanist-SemiBold.ttf</string>
+        <string>Urbanist-SemiBoldItalic.ttf</string>
+        <string>Urbanist-Thin.ttf</string>
+        <string>Urbanist-ThinItalic.ttf</string>
+    </array>
 </dict>
 </plist>
